requirements.txt
numpy>=1.24
Pillow>=10.0
h5py>=3.8
torch>=2.4
